Split test/compile/data.test.ts into test/compile/data/* #1641
Also, add more scaffolding + remove redundant tests.

Fix #1641
